# api/openapi.yaml
openapi: 3.1.0
info:
  title: NXT TIME PULSE Admin API
  version: 1.0.0
  description: Admin-gated API endpoints for GEO TIMESHEET - TIMEPULSE
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
paths:
...
  /assignments/{id}/end:
    post:
      summary: End assignment
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endDate]
              properties:
                endDate: { type: string, format: date }
                reason: { type: string }
      responses:
        "200":
          description: Assignment ended
        "202":
          description: Pending approval
...
  /admin/change-requests:
    get:
      summary: List change requests (admin approvals queue)
      tags: [Approvals]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected, cancelled] }
        - name: changeType
          in: query
          schema: { type: string, enum: [rate_change, assignment_change, user_access_change] }
      responses:
        "200":
          description: Change request list
  /admin/change-requests/{id}/approve:
    post:
      summary: Approve change request (reviewer must differ from requester)
      tags: [Approvals]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Approved
        "400":
          description: Invalid state or maker-checker violation
  /admin/change-requests/{id}/reject:
    post:
      summary: Reject change request
      tags: [Approvals]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Rejected
  /allocations:
    get:
      deprecated: true
      summary: Legacy compatibility endpoint (use /assignments)
      tags: [Compatibility]
      responses:
        "200":
          description: Compatibility payload
    post:
      deprecated: true
      summary: Legacy compatibility endpoint (use /assignments + /rates)
      tags: [Compatibility]
      responses:
        "200":
          description: Compatibility payload
        "202":
          description: Pending approval for critical/rate change
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...